完成回复功能，修复部分bug

+++ src/main/java/io/lgsity/qaforum/controller/PublishController.java
@@ -1,7 +1,6 @@
 package io.lgsity.qaforum.controller;
 
 import io.lgsity.qaforum.dto.QuestionDTO;
-import io.lgsity.qaforum.mapper.QuestionMapper;
 import io.lgsity.qaforum.pojo.Question;
 import io.lgsity.qaforum.pojo.User;
 import io.lgsity.qaforum.service.QuestionService;
@@ -35,7 +34,7 @@ public class PublishController {
     public String doPulish(@RequestParam(value = "title", required = false) String title,
                            @RequestParam(value = "description", required = false) String description,
                            @RequestParam(value = "tag", required = false) String tag,
-                           @RequestParam(value = "id", required = false) Integer id,
+                           @RequestParam(value = "id", required = false) Long id,
                            HttpServletRequest request,
                            Model model) {
         model.addAttribute("title", title);
@@ -70,7 +69,7 @@ public class PublishController {
     }
 
     @GetMapping("/publish/{id}")
-    public String edit(@PathVariable("id") Integer id,
+    public String edit(@PathVariable("id") Long id,
                        Model model){
         QuestionDTO question = questionService.getQuestionDTOByQuestionId(id);
         model.addAttribute("title", question.getTitle());
